Refactor test files
