+++ carbonmark-api/test/routes/categories/get.test.ts
@@ -1,9 +1,9 @@
 import { FastifyInstance } from "fastify";
 import nock from "nock";
-import { GRAPH_URLS } from "../../src/graphql/codegen.constants";
-import { build } from "../helper";
-import { DEV_URL } from "../test.constants";
-import { CATEGORIES, ERROR } from "./routes.mock";
+import { GRAPH_URLS } from "../../../src/graphql/codegen.constants";
+import { build } from "../../helper";
+import { DEV_URL } from "../../test.constants";
+import { CATEGORIES, ERROR } from "../routes.mock";
 
 describe("GET /categories", () => {
   let fastify: FastifyInstance;
