+++ carbonmark-api/test/routes/vintages/get.test.ts
@@ -1,42 +1,43 @@
 import { FastifyInstance } from "fastify";
 import nock from "nock";
-import { GRAPH_URLS } from "../../src/graphql/codegen.constants";
-import { build } from "../helper";
-import { DEV_URL } from "../test.constants";
-import { COUNTRIES, ERROR } from "./routes.mock";
+import { aProject } from "../../../src/.generated/mocks/marketplace.mocks";
+import { GRAPH_URLS } from "../../../src/graphql/codegen.constants";
+import { build } from "../../helper";
+import { DEV_URL } from "../../test.constants";
+import { ERROR } from "../routes.mock";
 
-describe("GET /countries", () => {
+describe("GET /vintages", () => {
   let fastify: FastifyInstance;
 
   // Setup the server
   afterEach(async () => await fastify.close());
   beforeEach(async () => {
     fastify = await build();
-    nock.cleanAll();
   });
 
   /** A default response for offsets */
   beforeEach(() =>
     nock(GRAPH_URLS.offsets)
       .post("")
-      .reply(200, { data: { carbonOffsets: COUNTRIES } })
+      .reply(200, { data: { carbonOffsets: [] } })
   );
 
   /** The happy path */
   test("Success", async () => {
+    const mock = aProject();
     nock(GRAPH_URLS.marketplace)
       .post("")
-      .reply(200, { data: { countries: COUNTRIES } });
+      .reply(200, { data: { projects: [mock] } });
 
     const response = await fastify.inject({
       method: "GET",
-      url: `${DEV_URL}/countries`,
+      url: `${DEV_URL}/vintages`,
     });
 
     const data = await response.json();
 
     expect(response.statusCode).toEqual(200);
-    expect(data).toEqual(COUNTRIES);
+    expect(data).toEqual([mock.vintage]);
   });
 
   /** An issue with one of the graph APIs */
@@ -49,7 +50,7 @@ describe("GET /countries", () => {
 
     const response = await fastify.inject({
       method: "GET",
-      url: `${DEV_URL}/countries`,
+      url: `${DEV_URL}/vintages`,
     });
 
     expect(response.body).toContain("Graph error occurred");
@@ -59,11 +60,11 @@ describe("GET /countries", () => {
   test("Empty data", async () => {
     nock(GRAPH_URLS.marketplace)
       .post("")
-      .reply(200, { data: { countries: [] } });
+      .reply(200, { data: { projects: [] } });
 
     const response = await fastify.inject({
       method: "GET",
-      url: `${DEV_URL}/countries`,
+      url: `${DEV_URL}/vintages`,
     });
 
     expect(response.statusCode).toBe(200);
@@ -74,11 +75,11 @@ describe("GET /countries", () => {
   test("Invalid data", async () => {
     nock(GRAPH_URLS.marketplace)
       .post("")
-      .reply(200, { data: { categories: "invalid data" } });
+      .reply(200, { data: { projects: "invalid data" } });
 
     const response = await fastify.inject({
       method: "GET",
-      url: `${DEV_URL}/countries`,
+      url: `${DEV_URL}/vintages`,
     });
 
     expect(response.statusCode).toEqual(502);
